// src/interfaces/member.ts
export interface Member {
    LIPSTARTYN: string;
    SEX: string;
    MEMID: number;
    EMAIL: string;
    NICKNAME: string;
    MEMBERSHIPYN: string;
    USE_YN: string;
    USERID: string;
    TEL: string;
    PWD: string;
    CRTDATE: string;
    BIRTHDAY: string;
    MEMNAME: string;
    MOBILE: string;
}